refactor into workspace
