+++ .gitignore
@@ -1,3 +1,3 @@
 /target
-.env
+**/.env
 .vscode
