+++ crates/cli/src/cmd.rs
