+++ crates/cli/src/main.rs
