+++ crates/server/.env.sample
@@ -18,11 +18,11 @@ AVALANCHE_RPC=
 # Do not require debug api and struct trace
 ETHEREUM_FORK_RPC=https://eth.llamarpc.com
 ARBITRUM_FORK_RPC=https://arb1.arbitrum.io/rpc
-# OPTIMISM_FORK_RPC=https://optimism.llamarpc.com
+OPTIMISM_FORK_RPC=https://optimism.llamarpc.com
 AVALANCHE_FORK_RPC=https://avalanche.public-rpc.com
 
 # Redis (optional)
-REDIS_ENABLED=1
+REDIS_ENABLED=0
 REDIS_HOST=localhost
 REDIS_PORT=9090
 REDIS_PASSWORD=
