+++ crates/token-bss/src/common.rs
