+++ crates/token-bss/src/ops/trace.rs
