+++ crates/token-bss/src/slot_finder.rs
