+++ crates/token-bss/src/trace_parser.rs
