--- src/slot_finder/mod.rs
@@ -1,10 +1,0 @@
-#[warn(clippy::module_inception)]
-mod trace_parser;
-mod slot_finder;
-mod lang;
-mod utils;
-mod ops;
-
-
-pub use slot_finder::find_balance_slots_and_update_ratio;
-pub use lang::EvmLanguage;
